start do confirm a recipe request from user to admin

+++ src/recipes/recipes.service.ts
@@ -2,12 +2,14 @@ import {
   Injectable,
   NotFoundException,
   InternalServerErrorException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { Recipe, RecipeDocument } from './schema/recipes.schema';
+import { Recipe, RecipeDocument, RecipeStatus } from './schema/recipes.schema';
 import { CreateRecipeDto } from './dto/create-recipe.dto';
 import { UpdateRecipeDto } from './dto/update-recipe.dto';
+import { UserRole } from 'src/users/schema/user.schema';
 
 @Injectable()
 export class RecipesService {
@@ -15,18 +17,32 @@ export class RecipesService {
     @InjectModel(Recipe.name) private recipeModel: Model<RecipeDocument>,
   ) {}
 
-  async create(createRecipeDto: CreateRecipeDto): Promise<Recipe> {
-    const createdRecipe = new this.recipeModel(createRecipeDto);
+  async create(
+    createRecipeDto: CreateRecipeDto,
+    authId: string,
+    isAdmin = false,
+  ): Promise<Recipe> {
+    const createdRecipe = new this.recipeModel({
+      ...createRecipeDto,
+      authId,
+      status: isAdmin ? RecipeStatus.APPROVED : RecipeStatus.PENDING,
+    });
     return await createdRecipe.save();
   }
 
   async findAll(): Promise<Recipe[]> {
     try {
-      return await this.recipeModel.find().exec();
+      return await this.recipeModel
+        .find({ status: RecipeStatus.APPROVED })
+        .exec();
     } catch (error) {
       console.error('Error in findall', error);
       return [];
     }
+  }
+
+  async findAllForAdmin(): Promise<Recipe[]> {
+    return this.recipeModel.find().exec();
   }
 
   async findOne(id: string): Promise<Recipe | null> {
@@ -42,8 +58,10 @@ export class RecipesService {
     }
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(id: string, userRole: UserRole): Promise<void> {
     try {
+      if (userRole !== UserRole.ADMIN)
+        throw new ForbiddenException('Only admins can delete recipes');
       const deleted = await this.recipeModel.findByIdAndDelete(id).exec();
       if (!deleted) {
         throw new NotFoundException(`Recipe with id ${id} not found`);
@@ -68,6 +86,20 @@ export class RecipesService {
     } catch (error) {
       console.error('Error updating recipe', error);
       throw new InternalServerErrorException('Could not update recipe');
+    }
+  }
+
+  async changeStatus(id: string, status: RecipeStatus): Promise<Recipe> {
+    try {
+      const updated = await this.recipeModel
+        .findByIdAndUpdate(id, { status }, { new: true })
+        .exec();
+      if (!updated)
+        throw new NotFoundException(`Recipe with id ${id} not found`);
+      return updated;
+    } catch (error) {
+      console.error('Error updating status recipe', error);
+      throw new InternalServerErrorException('Could not update status recipe');
     }
   }
 }
